+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
-* Breaking: change signature of Apps#create, move (some) params types to src/params
+* Breaking: move params types to src/params
+* Breaking: change signature of Apps#create, Addons#create
 * Breaking: compile destination is now `lib`
 * Breaking: move models (types and consts) to src/models